+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0-SNAPSHOT - 06/29/16
 
 ### Added
-
+- Create ParseObject and save to parse-server
 - Initial commit